refactor: 에러가 아닌 상황에도 로깅하도록 수정
