+++ src/main/java/in/koreatech/koin/global/exception/GlobalExceptionHandler.java
@@ -1,9 +1,6 @@
 package in.koreatech.koin.global.exception;
 
 import java.time.format.DateTimeParseException;
-import java.util.Enumeration;
-import java.util.HashMap;
-import java.util.Map;
 import java.util.stream.Collectors;
 
 import org.springframework.context.support.DefaultMessageSourceResolvable;
@@ -17,8 +14,6 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.context.request.ServletWebRequest;
 import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-import org.springframework.web.util.ContentCachingRequestWrapper;
-import org.springframework.web.util.WebUtils;
 
 import in.koreatech.koin.global.auth.exception.AuthenticationException;
 import in.koreatech.koin.global.auth.exception.AuthorizationException;
@@ -35,7 +30,6 @@ public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
         IllegalArgumentException e
     ) {
         log.warn(e.getMessage());
-        requestLogging(request);
         return buildErrorResponse(HttpStatus.BAD_REQUEST, e.getMessage());
     }
 
@@ -45,7 +39,6 @@ public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
         AuthorizationException e
     ) {
         log.warn(e.getMessage());
-        requestLogging(request);
         return buildErrorResponse(HttpStatus.FORBIDDEN, e.getMessage());
     }
 
@@ -55,7 +48,6 @@ public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
         AuthenticationException e
     ) {
         log.warn(e.getMessage());
-        requestLogging(request);
         return buildErrorResponse(HttpStatus.UNAUTHORIZED, e.getMessage());
     }
 
@@ -65,7 +57,6 @@ public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
         DataNotFoundException e
     ) {
         log.warn(e.getMessage());
-        requestLogging(request);
         return buildErrorResponse(HttpStatus.NOT_FOUND, e.getMessage());
     }
 
@@ -75,7 +66,6 @@ public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
         DuplicationException e
     ) {
         log.warn(e.getMessage());
-        requestLogging(request);
         return buildErrorResponse(HttpStatus.CONFLICT, e.getMessage());
     }
 
@@ -85,7 +75,6 @@ public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
         ExternalServiceException e
     ) {
         log.warn(e.getMessage());
-        requestLogging(request);
         return buildErrorResponse(HttpStatus.INTERNAL_SERVER_ERROR, e.getMessage());
     }
 
@@ -95,7 +84,6 @@ public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
         DateTimeParseException e
     ) {
         log.warn(e.getMessage());
-        requestLogging(request);
         return buildErrorResponse(HttpStatus.BAD_REQUEST, "잘못된 날짜 형식입니다. " + e.getParsedString());
     }
 
@@ -119,7 +107,6 @@ public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
             서버에서 에러가 발생했습니다. uri: {} {}
             {}
             """, request.getMethod(), request.getRequestURI(), detail);
-        requestLogging(request);
         return buildErrorResponse(HttpStatus.INTERNAL_SERVER_ERROR, "서버에서 오류가 발생했습니다.");
     }
 
@@ -144,7 +131,6 @@ public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
     ) {
         HttpServletRequest request = ((ServletWebRequest) webRequest).getRequest();
         log.warn("검증과정에서 문제가 발생했습니다. uri: {} {}, ", request.getMethod(), request.getRequestURI(), ex);
-        requestLogging(request);
         String errorMessages = ex.getBindingResult().getAllErrors().stream()
             .map(DefaultMessageSourceResolvable::getDefaultMessage)
             .collect(Collectors.joining(", "));
@@ -157,48 +143,5 @@ public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
     ) {
         var response = new ErrorResponse(httpStatus.value(), message);
         return ResponseEntity.status(httpStatus).body(response);
-    }
-
-    private void requestLogging(HttpServletRequest request) {
-        log.info("request header: {}", getHeaders(request));
-        log.info("request query string: {}", getQueryString(request));
-        log.info("request body: {}", getRequestBody(request));
-    }
-
-    private Map<String, Object> getHeaders(HttpServletRequest request) {
-        Map<String, Object> headerMap = new HashMap<>();
-        Enumeration<String> headerArray = request.getHeaderNames();
-        while (headerArray.hasMoreElements()) {
-            String headerName = headerArray.nextElement();
-            headerMap.put(headerName, request.getHeader(headerName));
-        }
-        return headerMap;
-    }
-
-    private String getQueryString(HttpServletRequest httpRequest) {
-        String queryString = httpRequest.getQueryString();
-        if (queryString == null) {
-            return " - ";
-        }
-        return queryString;
-    }
-
-    private String getRequestBody(HttpServletRequest request) {
-        var wrapper = WebUtils.getNativeRequest(request, ContentCachingRequestWrapper.class);
-        if (wrapper == null) {
-            return " - ";
-        }
-        try {
-            // body 가 읽히지 않고 예외처리 되는 경우에 캐시하기 위함
-            wrapper.getInputStream().readAllBytes();
-            byte[] buf = wrapper.getContentAsByteArray();
-            if (buf.length == 0) {
-                return " - ";
-            }
-            return new String(buf, wrapper.getCharacterEncoding());
-        } catch (Exception e
-        ) {
-            return " - ";
-        }
     }
 }
